setup_requirements.txt
flask==3.0.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
PyPDF2==3.0.1
openai==1.3.5
python-dotenv==1.0.0

# Authentication and security
oauthlib==3.2.2
google-auth==2.23.4

# Development tools (optional)
pytest==7.4.3
black==23.11.0